+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "node tester.js"
+    "test": "node test.js"
   },
   "repository": {
     "type": "git",
